+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./agent";
 export * from "./ids";
 export * from "./operations";
 export * from "./shapes";